Fix extracting cookies from Chrome >= 130
Chrome 130 changed the format of the encrypted_value for the cookies. It
now adds the SHA256 hash of the domain name to the start of the
encrypted data, so we need to remove this.

See https://github.com/chromium/chromium/commit/5ea6d65c622a3d5ff75db9dc0257ea3869f31289
for more details.