Use system Python version in build job in CI
Using setup-python may cause it to setup things so WeeChat will use the
Python packages for that version of Python instead of the system level
Python packages, so it won't find websocket from python3-websocket.

WeeChat will run with the system version of Python regardless, which is
why we want it to use the packages for that version, rather than
installing websocket-client in the environment setup-python sets up.